[cascade.executor] switch worker from fork to forkserver
as demonstrated on a torch anemoi job with XPC_ERROR_CONNECTION_INVALID,
we cant fork worker processes, at least not on macos. Possibly due
to the gateway process have been tainted somehow -- but not sure how,
I dont think there is any torch import happening prior to work start

this makes the non-persistent worker scenario much less performant.
If we ever want to go that way again, we'd need to create a dedicated
process-factory process or re-investigate this again

this in turn required two minor fixes:
- move the benchmarks/__main__ code into its own module. This should
  have been done anyway
- guard runnerContext with cloudpickle due to lambdas usage. Alt we
  could have gotten rid of the defaultdicts, but that would be prolly
  uglier

in other situations (shm/dserver process starts), we retain forks for
now, as it does not seem to be a problem
